refactor: move get_current_user to api/dependencies to fix layer violation
- foundation/auth/dependencies.py now only contains parse_user_id_header
  (no DB/repo dependency)
- get_current_user, require_admin, get_user_repository moved to
  api/dependencies.py (correct layer for DI assembly)
- All routers updated to import from api.dependencies
- Remove unused user_id from UpdateMyProfileInput DTO
