+++ backend/contexts/record/presentation/router.py
@@ -134,7 +134,7 @@ from contexts.record.presentation.schemas import (
     UpdateMemoRequest,
     UpdateMemoResponse,
 )
-from foundation.auth.dependencies import get_current_user
+from api.dependencies import get_current_user
 from shared.domain.user import User
 from shared.domain.value_objects import UserId
