+++ backend/foundation/auth/dependencies.py
@@ -1,22 +1,22 @@
 from __future__ import annotations
 
 import uuid
-from typing import Annotated
 
-from fastapi import Depends, Header, HTTPException
+from fastapi import Header, HTTPException
 
-from shared.domain.user import User
-from shared.domain.user_repository import UserRepository
 from shared.domain.value_objects import UserId
-from shared.presentation.dependencies import get_user_repository
 
 _ERROR_DETAIL = "X-User-Id header is missing or invalid"
 
 
-def _parse_user_id_header(
+def parse_user_id_header(
     x_user_id: str | None = Header(default=None),
 ) -> UserId:
-    """Parse and validate the X-User-Id header value."""
+    """Parse and validate the X-User-Id header value.
+
+    This is a development-only header parser.
+    Replace with JWT token extraction for production use.
+    """
     if x_user_id is None:
         raise HTTPException(status_code=401, detail=_ERROR_DETAIL)
 
@@ -26,37 +26,3 @@ def _parse_user_id_header(
         raise HTTPException(
             status_code=401, detail=_ERROR_DETAIL,
         ) from None
-
-
-async def get_current_user(
-    parsed_id: Annotated[UserId, Depends(_parse_user_id_header)],
-    user_repo: Annotated[UserRepository, Depends(get_user_repository)],
-) -> User:
-    """Extract and verify User from X-User-Id header.
-
-    This is a development-only authentication dependency.
-    Replace the implementation body with JWT validation (or
-    similar) for production use -- router signatures unchanged.
-    """
-    user = await user_repo.get_by_id(parsed_id)
-    if user is None:
-        raise HTTPException(
-            status_code=403, detail="User not found",
-        )
-    if not user.is_active:
-        raise HTTPException(
-            status_code=403,
-            detail="User account is deactivated",
-        )
-    return user
-
-
-async def require_admin(
-    current_user: Annotated[User, Depends(get_current_user)],
-) -> User:
-    """Ensure the current user has admin role."""
-    if not current_user.is_admin:
-        raise HTTPException(
-            status_code=403, detail="Admin access required",
-        )
-    return current_user
