+++ backend/tests/contexts/record/presentation/test_router.py
@@ -98,7 +98,7 @@ from contexts.record.presentation.dependencies import (
     get_update_memo_use_case,
 )
 from contexts.record.presentation.router import router
-from foundation.auth.dependencies import get_current_user
+from api.dependencies import get_current_user
 from shared.domain.user import User
 from shared.domain.value_objects import UserId, UserRole
